added census of India 2011 village-town list
Note on naming convention
[link](http://censusindia.gov.in/2011census/Listofvillagesandtowns.aspx)
As per the recommendations of the MDDS Committee a new coding pattern
for various geographical entities has been devised. The Location Code
Directory has been provided unique codes on All India basis which can
be used for various e-Governance purposes. The coding convention used
is as described below:

1. The State code of 2 digits within India has been used similar to
the one used in earlier census.

2. The District code of 3 digits continuous code within India has been
given.

3. The Sub-district code of 5 digits continuous code within India has
been used.

4. The Village code of 6 digits has been used continuous code within
India. Range of 000001 � 799999 has been kept for Villages. The Census
Towns and OGs which are treated as urban for census purpose have also
been included in the Rural Directory. The town code (only statutory
towns) have been numbered within India. The town code range is 800001
- 899999.

Further, the provision for new addition, merger or split cases shall
be as follows:

Addition Cases:

In case of addition of new Villages/Towns/Sub-DTs/DTs, the new
geographical entity shall be added at the end with an incremental code
for that geographical entity.

Merger Cases:

In case of merger, the original code of merging geographical units
shall be frozen and a new single incremental code shall be given to
the merged unit at the end of the corresponding geographical entity.

Split Cases:

In case of split cases the original single code shall be frozen and
new individual incremental codes shall be provided at the end of the
corresponding geographical entity.